+++ R/boilerplate-cubist.R
@@ -117,9 +117,9 @@ hai_auto_cubist <- function(.data, .rec_obj, .splits_obj = NULL, .rsamp_obj = NU
   }
 
   # Model Specification ----
-  model_spec <- model_spec %>%
-    parsnip::set_mode(mode = model_type) %>%
-    parsnip::set_engine(engine = "Cubist")
+  #model_spec <- model_spec %>%
+  #  parsnip::set_mode(mode = model_type) %>%
+  #  parsnip::set_engine(engine = "Cubist")
 
   # Workflow ----
   wflw <- workflows::workflow() %>%
@@ -145,8 +145,8 @@ hai_auto_cubist <- function(.data, .rec_obj, .splits_obj = NULL, .rsamp_obj = NU
     }
 
     # Tune the workflow
-    # Start parallel backed
-    modeltime::parallel_start(num_cores)
+    # Start parallel backend
+    cl <- parallelly::makeClusterPSOCK(workers = num_cores)
 
     tuned_results <- wflw %>%
       tune::tune_grid(
@@ -155,7 +155,7 @@ hai_auto_cubist <- function(.data, .rec_obj, .splits_obj = NULL, .rsamp_obj = NU
         metrics   = ms
       )
 
-    modeltime::parallel_stop()
+    parallel::stopCluster(cl)
 
     # Get the best result set by a specified metric
     best_result_set <- tuned_results %>%